refactor: clean up the CLI tool
