+++ example/src/index.scss
@@ -1,4 +1,4 @@
-@import "@edx/brand-openedx/paragon/fonts";
-@import "@edx/brand-openedx/paragon/variables";
+@import "@edx/brand/paragon/fonts";
+@import "@edx/brand/paragon/variables";
 @import "~@edx/paragon/scss/core/core";
-@import "@edx/brand-openedx/paragon/overrides";
+@import "@edx/brand/paragon/overrides";
